requirements.txt
numpy==1.26.4
face_recognition==1.3.0
opencv-python==4.9.0.80
psutil==5.9.8
customtkinter==5.2.2
google-generativeai==0.5.4
pyautogui==0.9.54
cvlib==0.2.7
pyttsx3==2.90
SpeechRecognition==3.10
pycaw==20230322
pywhatkit==5.6.1
pyperclip==1.8.2
pyowm==3.3.0
pyjokes==0.6.0
pyaudio==0.2.13 
tensorflow==2.19.0
requests==2.31.0
requests-html==0.10.0
beautifulsoup4==4.12.2
speedtest-cli==2.1.3
wmi==1.5.1
pywhatkit==5.4